+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Unreleased
 ----------
 - Renamed `api::market_status::Status` type to `Market`
 - Introduced `Status` type representing market status
+- Fixed parsing of market time to assume server-local time
 
 
 0.9.0